Calculates physical crop area at the subnational level

To estimate the physical crop area for each farming system harvest area (ha) statistics are combined with information on farmings system shares (fs) and cropping intensity (ci). Depending on how the model is solved, the physical area statistics are saved at the administrative unit level 0 (country) level (model solve level 0) or at the level 1 administrative unit level (model solve level 1).

prepare_physical_area(param)

Arguments

param

Object of type mapspamc_par that bundles all mapspamc parameters, including core model folders, alpha-3 country code, year, spatial resolution, most detailed level at which subnational statistics are available, administrative unit level at which the model is solved and type of model.

Details

prepare_physical_area combines ha, fs and ci statistics and saves two files in csv format: (1) physical area (pa) and (2) physical area broken down by farming systems (pa_fs). Results are saved in the subfolders that are located in the the processed_data/intermediate folder.

Examples

if (FALSE) {
prepare_physical_area(param)
}