Creates template for subnational crop statistics

To support the preparation of the subnational statistics, create_statistics_template() can create three types of data templates:

ha for harvest area statistics

fs for farming system share

ci for cropping intensity.

create_statistics_template(type, param)

Arguments

type

Character vector that refers to the type of template that needs to be created. See details for allowed input.

Details

The function requires information on how the different administrative unit levels are nested. This file needs to be created first by running create_adm_list().

The dimensions of the templates are determined by the parameters set by means of mapspamc_par(). The ha template dimension is determined by adm level parameter, while the dimensions of the fs and cs templates (which are identical) are determined by the solve level parameter.

Examples

create_statistics_template(type = "ha", param)
#> Error in stopifnot(inherits(param, "mapspamc_par")): object 'param' not found